% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl.R
\name{ncl_baseurl}
\alias{ncl_baseurl}
\title{Builds the Nextcloud base URL for a nextcloud server from host and user names}
\usage{
ncl_baseurl(hostname, username, path_prefix = "")
}
\arguments{
\item{hostname}{host name}

\item{username}{user name}

\item{path_prefix}{optional path to webdav directory}
}
\value{
base url to use with \code{\link{wd_connect}}
}
\description{
Builds the Nextcloud base URL for a nextcloud server from host and user names
}
\examples{
ncl_baseurl("example.com","johndoe")
ncl_baseurl("example.com","johndoe","sub/dir/")

}
