% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl.R
\name{ncl_shareurl}
\alias{ncl_shareurl}
\title{Builds the Nextcloud base URL for a public share from host and user names}
\usage{
ncl_shareurl(hostname, username, path_prefix = "")
}
\arguments{
\item{hostname}{host name}

\item{username}{user name}

\item{path_prefix}{ptional path to webdav directory}
}
\value{
share url to use with \code{\link{wd_connect}}
}
\description{
Builds the Nextcloud base URL for a public share from host and user names
}
\examples{
ncl_shareurl("example.com", "johndoe")

}
