% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl.R
\name{ncl_shareurl_from_publicurl}
\alias{ncl_shareurl_from_publicurl}
\title{Creates the Nextcloud base URL for a share from a public share url}
\usage{
ncl_shareurl_from_publicurl(url)
}
\arguments{
\item{url}{link to public share}
}
\value{
share url to use with \code{\link{wd_connect}}
}
\description{
Creates the Nextcloud base URL for a share from a public share url
}
\examples{
ncl_shareurl_from_publicurl("https://example.com/s/87d7edad/")

}
