% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncl.R
\name{ncl_username_from_url}
\alias{ncl_username_from_url}
\title{Extracts the user name from a Nextcloud Base URL}
\usage{
ncl_username_from_url(url)
}
\arguments{
\item{url}{base or share url}
}
\value{
user name
}
\description{
Extracts the user name from a Nextcloud Base URL
}
\examples{
ncl_username_from_url("https://example.com/remote.php/dav/files/johndoe")

}
