% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs.R
\name{ocs_delete_share}
\alias{ocs_delete_share}
\title{Deletes a share}
\usage{
ocs_delete_share(req, id)
}
\arguments{
\item{req}{WebDAV request as returned by \code{\link{wd_connect}}}

\item{id}{share id}
}
\value{
invisible boolean TRUE on success and FALSE on failure
}
\description{
Deletes a share
}
\examples{
\dontrun{
r <- wd_connect("https://example.com/remote.php/dav/files/johndoe")
ocs_delete_share(r, 12342)
}

}
