% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs.R
\name{ocs_send_mail}
\alias{ocs_send_mail}
\title{Notifies the user of a mail share}
\usage{
ocs_send_mail(req, id, password = NULL)
}
\arguments{
\item{req}{WebDAV request as returned by \code{\link{wd_connect}}}

\item{id}{share id}

\item{password}{password of the share if it is password protected}
}
\value{
invisible TRUE on success or FALSE on failure
}
\description{
Notifies the user of a mail share
}
\examples{
\dontrun{
r <- wd_connect("https://example.com/remote.php/dav/files/johndoe")

# add a password to a mail share and notify the user
ocs_modify_share(r, 12342, password = "super_secret")
ocs_send_mail(r, 12342, password = "super_secret")
}
}
