% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocs.R
\name{ocs_share_info}
\alias{ocs_share_info}
\title{Get infor for a specific share}
\usage{
ocs_share_info(req, id, columns = NULL)
}
\arguments{
\item{req}{WebDAV request as returned by \code{\link{wd_connect}}}

\item{id}{share id}

\item{columns}{column names that should be included into the result (default \code{NULL} includes all)}
}
\value{
one row data.frame with share properties
}
\description{
Get infor for a specific share
}
\examples{
\dontrun{
r <- wd_connect("https://example.com/remote.php/dav/files/johndoe")
ocs_share_info(r, 158742)
}
}
