% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_delete}
\alias{wd_delete}
\title{Deletes a file or directory (collection) on WebDAV server}
\usage{
wd_delete(req, file)
}
\arguments{
\item{req}{request handle obtained from \code{\link{wd_connect}}}

\item{file}{path to file or directory to delete on the server}
}
\value{
TRUE on success, FALSE on failure (invisibly)
}
\description{
Deletes a file or directory (collection) on WebDAV server
}
\examples{
\dontrun{

wd_delete(r, "testfile.R")

}
}
