% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_getwd}
\alias{wd_getwd}
\title{Get the current WebDAV working directory}
\usage{
wd_getwd(req)
}
\arguments{
\item{req}{request handle obtained from \code{\link{wd_connect}}}
}
\value{
name of the WebDAV directory
}
\description{
Get the current WebDAV working directory
}
\examples{
\dontrun{
wd_getwd(req)
}
}
