% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_isdir}
\alias{wd_isdir}
\title{Checks if the resource on WebDAV is a directory}
\usage{
wd_isdir(req, directory, silent = FALSE)
}
\arguments{
\item{req}{request handle obtained from \code{\link{wd_connect}}}

\item{directory}{path to directory}

\item{silent}{if FALSE a warning is given if the directory does not exists}
}
\value{
TRUE if it is a directory, FALSE else
}
\description{
Checks if the resource on WebDAV is a directory
}
\examples{
\dontrun{

wd_isdir(r, "testfile.R") # FALSE
wd_isdir(r, "mydirectory")   # TRUE

}
}
