% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_mkdir}
\alias{wd_mkdir}
\title{Creates a directory (collection) on WebDAV server}
\usage{
wd_mkdir(req, directory)
}
\arguments{
\item{req}{request handle obtained from \code{\link{wd_connect}}}

\item{directory}{directory path on server}
}
\value{
TRUE on success, FALSE on failure (invisibly)
}
\description{
When creating a subdirectoy, all parent directories have to exist on the
server.
}
\examples{
\dontrun{

# creates 'newdir' inside the subdirectory 'existing/directory'
wd_mkdir(r, "existing/directory/newdir")

}
}
