% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_move}
\alias{wd_move}
\title{Moves a file or directory on the server}
\usage{
wd_move(req, source, target, overwrite = TRUE)
}
\arguments{
\item{req}{request handle obtained from \code{\link{wd_connect}}}

\item{source}{path of the source on the server}

\item{target}{path of the target on the server}

\item{overwrite}{overwrites files  when TRUE (default)}
}
\value{
TRUE on success, FALSE on failure (invisibly)
}
\description{
Moves a file or directory on the server
}
\examples{
\dontrun{

wd_move(r, "testfile.R", "testfile_old.R")

}
}
