% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_connect}
\alias{wd_connect}
\title{Establishes a connection to a WebDAV server}
\usage{
wd_connect(url, user)
}
\arguments{
\item{url}{url of the webdav folder}

\item{user}{username}
}
\value{
a httr2 request
}
\description{
Creates and authenticate a request handle to the webserver
}
\examples{
  \dontrun{
# establish a connection, you will be asked for a password

r <- wd_connect("https://example.com/remote.php/webdav/","exampleuser")
  }
}
