% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_copy}
\alias{wd_copy}
\title{Copies a file or folder on the WebDAV server}
\usage{
wd_copy(req, source, target)
}
\arguments{
\item{req}{request handle}

\item{source}{path of the source on server}

\item{target}{target path on the server}
}
\value{
invisibly true on success or false on failure
}
\description{
Copies a file or folder on the WebDAV server
}
