% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_delete}
\alias{wd_delete}
\title{Deletes a file or folder (collection) on WebDAV server}
\usage{
wd_delete(req, folder)
}
\arguments{
\item{req}{request handle}

\item{folder}{folder}
}
\value{
invisibly true on success or false on failure
}
\description{
Deletes a file or folder (collection) on WebDAV server
}
\examples{
\dontrun{

wd_delete(r, "testfile.R")

}
}
