% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_dir}
\alias{wd_dir}
\title{Lists the content of a WebDAV directory}
\usage{
wd_dir(req, folder = "", full_names = FALSE, as_df = FALSE)
}
\arguments{
\item{req}{request handle}

\item{folder}{folder path}

\item{full_names}{whether filenames should be returned with path}

\item{as_df}{outputs a data.frame with file information}
}
\value{
a vector of filenames or a dataframe
}
\description{
Lists the content of a WebDAV directory
}
