% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_download}
\alias{wd_download}
\title{Fetches a file or folder (recursively) from the WebDAV server}
\usage{
wd_download(req, source, target = "")
}
\arguments{
\item{req}{request handle}

\item{source}{source file or folder on server}

\item{target}{local target folder}
}
\value{
vector of downloaded files
}
\description{
Fetches a file or folder (recursively) from the WebDAV server
}
\examples{
  \dontrun{

wd_download(r, "weatherfiles", "d:/data/weather")
wd_download(r, "test/xyz.txt", "d:/data/abc.txt")

  }
}
