% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_isdir}
\alias{wd_isdir}
\title{Checks if the resource on WebDAV is a folder}
\usage{
wd_isdir(req, folder, silent = FALSE)
}
\arguments{
\item{req}{request handle}

\item{folder}{path to folder}

\item{silent}{if FALSE a warning is given if the folder does not exists}
}
\value{
TRUE if it is a folder, FALSE else
}
\description{
Checks if the resource on WebDAV is a folder
}
\examples{
  \dontrun{

wd_isdir(r, "testfile.R") # FALSE
wd_isdir(r, "myfolder")   # TRUE

  }
}
