% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_isdir}
\alias{wd_isdir}
\title{Checks if the resource on WebDAV is a folder}
\usage{
wd_isdir(req, folder)
}
\arguments{
\item{req}{request handle}

\item{folder}{path to folder}
}
\value{
TRUE if it is a folder, FALSE else
}
\description{
Checks if the resource on WebDAV is a folder
}
