% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_mkdir}
\alias{wd_mkdir}
\title{Creates a directory (collection) on WebDAV server}
\usage{
wd_mkdir(req, folder)
}
\arguments{
\item{req}{request handle}

\item{folder}{folder path on server}
}
\value{
@return invisibly true on success or false on failure
}
\description{
Creates a directory (collection) on WebDAV server
}
