% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_move}
\alias{wd_move}
\title{Moves a file or folder on the server}
\usage{
wd_move(req, source, target)
}
\arguments{
\item{req}{request handle}

\item{source}{path of the source on server}

\item{target}{target path on the server}
}
\value{
invisibly true on success or false on failure
}
\description{
Moves a file or folder on the server
}
\examples{
  \dontrun{

wd_move(r, "testfile.R", "testfile_old.R")

  }
}
