% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdav.R
\name{wd_upload}
\alias{wd_upload}
\title{Uploads a file or folder to WebDAV}
\usage{
wd_upload(req, source, target = "")
}
\arguments{
\item{req}{request handle}

\item{source}{local file or folder}

\item{target}{remote file or foldername}
}
\value{
vector of uploaded files
}
\description{
Uploads a file or folder to WebDAV
}
\examples{
  \dontrun{

wd_upload(r, "d:/data/weather", "weatherfiles")
wd_upload(r, "d:/data/abc.txt", "test/xyz.txt")

  }
}
